Implements Chef DTO and data transfer
Introduces a Chef DTO to decouple the API from the entity model.
This change allows for more flexible data transfer and avoids exposing
internal data structures directly. Also, adds a mapper for conversion
between the entity and DTO. It enables the front end to send requests
to the backend using the correct CORS configuration.